Renamed Directory screens in edit_flashcard_use_case to function_testing.
